refactor: move CommandExecutor and command rendering into monochange_publish
